fix: support ESM for indexTransform
