+++ examples/custom-webpack/full-cycle-app/package.json
@@ -32,7 +32,7 @@
     "@angular/language-service": "13.2.0",
     "@types/jasmine": "3.10.3",
     "@types/node": "16.11.21",
-    "cypress": "latest",
+    "cypress": "9.3.1",
     "html-webpack-plugin": "5.5.0",
     "jasmine-core": "4.0.0",
     "karma": "6.3.12",
