+++ examples/custom-webpack/sanity-app-esm/angular.json
@@ -111,7 +111,7 @@
             "main": "src/test.ts",
             "polyfills": "src/polyfills.ts",
             "tsConfig": "tsconfig.spec.json",
-            "karmaConfig": "karma.conf.js",
+            "karmaConfig": "karma.conf.cjs",
             "assets": [
               "src/favicon.ico",
               "src/assets"
