+++ examples/custom-webpack/sanity-app-esm/package.json
@@ -41,6 +41,7 @@
     "karma-jasmine": "4.0.1",
     "karma-jasmine-html-reporter": "1.7.0",
     "typescript": "4.5.5",
-    "cypress": "latest"
+    "cypress": "9.3.1",
+    "puppeteer": "13.1.2"
   }
 }
