+++ examples/jest/simple-app/package.json
@@ -32,7 +32,7 @@
     "@angular/language-service": "13.2.0",
     "@types/jasmine": "3.10.3",
     "@types/node": "16.11.21",
-    "cypress": "latest",
+    "cypress": "9.3.1",
     "jasmine-core": "4.0.0",
     "jest": "27.4.7",
     "jest-junit": "13.0.0",
