+++ examples/timestamp/package.json
@@ -34,7 +34,7 @@
     "@cypress/schematic": "1.6.0",
     "@types/jasmine": "3.10.3",
     "@types/node": "16.11.21",
-    "cypress": "latest",
+    "cypress": "9.3.1",
     "jasmine-core": "4.0.0",
     "karma": "6.3.12",
     "karma-chrome-launcher": "3.1.0",
